+++ lib/slack/real_time/concurrency/celluloid.rb
@@ -68,9 +68,12 @@ module Slack
           end
 
           def start_async(client)
-            Thread.new do
-              client.run_loop
-            end
+            @client = client
+            Actor.new(future.run_client_loop)
+          end
+
+          def run_client_loop
+            @client.run_loop
           end
 
           def connected?
@@ -78,6 +81,18 @@ module Slack
           end
 
           protected
+
+          class Actor
+            attr_reader :future
+
+            def initialize(future)
+              @future = future
+            end
+
+            def join
+              @future.value
+            end
+          end
 
           def build_socket
             socket = ::Celluloid::IO::TCPSocket.new(addr, port)